agent/requirements.txt
anthropic>=0.40.0
langfuse>=2.36.0,<3.0.0
prometheus-client>=0.21.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
insightface>=0.7.3
onnxruntime>=1.20.0
opencv-python-headless>=4.10.0
numpy>=1.26.0
pillow>=10.4.0
scipy>=1.13.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
structlog>=24.4.0
